Added some new behavioral testing and defaults
Set up the environment so that it mimics the default inputs that Azure DevOps
will provide. This means the vulnerabilities.json had to be mocked so the info
could be printed out by default.

Also added in some testing around the failBuild behavior where it checks the
result of the policy scan and uses that along with the failBuild parameter to
determine whether the pipeline should fail.